Refactored notify_osd
